+++ src/practices/Java/JavaPackageManagementUsedPractice.ts
@@ -9,7 +9,7 @@ import { PracticeContext } from '../../contexts/practice/PracticeContext';
   impact: PracticeImpact.high,
   suggestion: 'Use pom.xml or build.gradle to keep track of packages that are being used in your application.',
   reportOnlyOnce: true,
-  url: 'https://maven.apache.org/',
+  url: 'https://dxkb.io/p/package-management',
 })
 export class JavaPackageManagementUsedPractice implements IPractice {
   async isApplicable(ctx: PracticeContext): Promise<boolean> {
@@ -24,6 +24,8 @@ export class JavaPackageManagementUsedPractice implements IPractice {
     if (await ctx.fileInspector.exists('pom.xml')) {
       return PracticeEvaluationResult.practicing;
     } else if (await ctx.fileInspector.exists('build.gradle')) {
+      return PracticeEvaluationResult.practicing;
+    } else if (await ctx.fileInspector.exists('build.gradle.kts')) {
       return PracticeEvaluationResult.practicing;
     }